pyproject.toml
[project]
name = "cqi-be-main"
version = "0.1.0"
description = "Code Quality Intelligence Backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core web framework
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic>=2.7.4",
    "python-multipart==0.0.6",

    # Essential AI packages
    "langchain-groq>=0.1.9",
    "langchain-core>=0.2.26",
    "langgraph>=0.1.1",
    "langsmith>=0.0.87",

    # Environment utilities
    "python-dotenv==1.0.0",

    # Git integration
    "GitPython==3.1.40",
    "PyGithub==1.59.1",

    # ML dependencies
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",

    # Additional frameworks
    "flask>=2.3.0",
]
